requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
cachetools==5.3.3
python-multipart==0.0.9